// src/_services/api/model/resultTransporteDtoOutputConferencia.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Gestão de Devolução e Produtividade
 * Documentação da API responsável por: gestão de produtividade, montagem e impressão de mapas de separação, sistema de devolução de estoque e contagem de inventário.
 * OpenAPI spec version: 1.0
 */

export type ResultTransporteDtoOutputConferencia =
  (typeof ResultTransporteDtoOutputConferencia)[keyof typeof ResultTransporteDtoOutputConferencia];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ResultTransporteDtoOutputConferencia = {
  NAO_INICIADO: 'NAO_INICIADO',
  EM_PROGRESSO: 'EM_PROGRESSO',
  CONCLUIDO: 'CONCLUIDO',
  EM_PAUSA: 'EM_PAUSA',
} as const;